Support restart of downstairs during repair (#300)
This adds the ability for the upstairs to repair mismatching  downstairs
even if they go offline during the repair.  The repair may start over, or
it may continue from where it left off if some extents have been repaired.

upstairs/lib.rs
During recovery, we now handle a downstairs returning error and this
will restart the entire recovery (all downstairs will reconnect).
The upstairs now checks a little more frequently for a missing
downstairs.
Retry activation forever (as repair is now part of activation).
Update message when a read hash does not match between downstairs.
Update message when we get a read error from a downstairs.

upstairs/mend.rs
More messages about what we are repairing.

client.rs
Added range support to the validate_vec() function if the caller desires.
This enables buffer validation from a range of expected write counts for
a block, a condition that could exist after a recovery when IO was in
flight.
Updated the repair workload to support the range option.  This enables
the test to commit the write log when we have a flush and we should
expect any future repairs to have a write count in the range recorded.
Added range support to the validate workload, so we can recover from
the repair workload.

cli.rs:
Updates for supporting the validate_vec with range option.

common/lib.rs
New error types for failures during ds->ds repair

downstairs/dump.rs
Flexible column width printing to support larger physical offsets.A

downstairs/lib.rs
Print a message when we return error due to wrong uuid

downstairs/region.rs
No more panics when repair things are not as we expect.  We now do a
better job when we get an error (missing file, broken stream) and
report that error back to the upstairs as a failed repair job.

fsync fsync for everyone.  fsync all new files after we strem in the
data for them.  fsync files after moving.  fsync directories after
moving, and fsync containing directories after moving files.

downstairs/repair.rs
Return better errors, issue #289

tools/create-generic-ds.sh
Better help, more messaging, time output for downstairs create.

tools/test_repair.sh
Cleanup and improve the test to better handle failures, to restart
any downstairs, to support the new --range option, and to give me
more messages when it does fail so I can figure out what broke.

tools/test_restart_repair.sh
New test to verify repair can work while downstairs are restarting.
This test tests the changes made in this commit.

tools/test_up.sh
Put a dump test in the middle of the repair test, we can now check
both the repair test has introduced a difference, and that the dump
command can detect it.

tools/downstairs-daemon.sh
internal self cleanup to avoid getting in my own way.

+++ tools/create-generic-ds.sh
@@ -6,22 +6,23 @@ ROOT=$(cd "$(dirname "$0")/.." && pwd)
 
 cd "$ROOT" || (echo failed to cd "$ROOT"; exit 1)
 
-if pgrep -fl crucible-downstairs; then
-    echo 'Some downstairs already running?' >&2
-    exit 1
-fi
-
-if ! cargo build; then
-    echo "Initial Build failed, no tests ran"
-    exit 1
-fi
+usage () {
+    echo "Usage: $0 [de] [-b #] [-c #] [-s #]" >&2
+    echo " -b block_size    Block size for the region" >&2
+    echo " -c extent_count  Total number of extent files" >&2
+    echo " -d               Delete existing var/88* direcories" >&2
+    echo " -e               Require encryption on the volume" >&2
+    echo " -h               Show this help message" >&2
+    echo " -s extent_size   Number of extents per extent file" >&2
+}
 
 block_size=512
 delete=0
 encryption=0
 extent_size=100
 extent_count=20
-while getopts 'b:c:des:' opt; do
+
+while getopts 'b:c:dehs:' opt; do
     case "$opt" in
         b)  block_size=$OPTARG
             echo "Using block size $block_size"
@@ -35,19 +36,28 @@ while getopts 'b:c:des:' opt; do
         e)  encryption=1
             echo "enable encryption"
             ;;
+        h)  usage
+            exit 0
+            ;;
         s)  extent_size=$OPTARG
             echo "Using extent size $extent_size"
             ;;
-        *)  echo "Usage: $0 [de] [-b #] [-c #] [-s #]" >&2
-            echo " -b block_size    Block size for the region" >&2
-            echo " -c extent_count  Total number of extent files" >&2
-            echo " -d                Delete existing var/88* direcories" >&2
-            echo " -e                Require encryption on the volume" >&2
-            echo " -s extent_size   Number of extents per extent file" >&2
+        *)  echo "Invalid option"
+            usage
             exit 1
             ;;
     esac
 done
+
+if pgrep -fl crucible-downstairs; then
+    echo 'Some downstairs already running?' >&2
+    exit 1
+fi
+
+if ! cargo build; then
+    echo "Initial Build failed, no tests ran"
+    exit 1
+fi
 
 if [[ $delete -eq 1 ]]; then
     rm -rf var/8810 var/8820 var/8830
@@ -58,9 +68,9 @@ else
     fi
 fi
 
-cds="./target/debug/crucible-downstairs"
+cds="./target/release/crucible-downstairs"
 if [[ ! -f ${cds} ]]; then
-    cds="./target/release/crucible-downstairs"
+    cds="./target/debug/crucible-downstairs"
     if [[ ! -f ${cds} ]]; then
         echo "Can't find crucible binary at $cds"
         exit 1
@@ -70,16 +80,17 @@ fi
 res=0
 for port in 8810 8820 8830; do
     if [[ $encryption -eq 0 ]]; then
-        if ! cargo run -q -p crucible-downstairs -- create -u 12345678-"$port"-"$port"-"$port"-00000000"$port" -d var/"$port" --extent-count "$extent_count" --extent-size "$extent_size" --block-size "$block_size"; then
+        echo cargo run -q -p crucible-downstairs -- create -u 12345678-"$port"-"$port"-"$port"-00000000"$port" -d var/"$port" --extent-count "$extent_count" --extent-size "$extent_size" --block-size "$block_size"
+        if ! time "$cds" create -u 12345678-"$port"-"$port"-"$port"-00000000"$port" -d var/"$port" --extent-count "$extent_count" --extent-size "$extent_size" --block-size "$block_size"; then
             echo "Failed to create downstairs $port"
             res=1
         fi
     else
-        if ! cargo run -q -p crucible-downstairs -- create -u 12345678-"$port"-"$port"-"$port"-00000000"$port" -d var/"$port" --extent-count "$extent_count" --extent-size "$extent_size"  --block-size "$block_size" --encrypted=true; then
+        echo cargo run -q -p crucible-downstairs -- create -u 12345678-"$port"-"$port"-"$port"-00000000"$port" -d var/"$port" --extent-count "$extent_count" --extent-size "$extent_size"  --block-size "$block_size" --encrypted=true
+        if ! time "$cds" create -u 12345678-"$port"-"$port"-"$port"-00000000"$port" -d var/"$port" --extent-count "$extent_count" --extent-size "$extent_size"  --block-size "$block_size" --encrypted=true; then
             echo "Failed to create downstairs $port"
             res=1
         fi
-
     fi
 done
 exit $res
